move any axios stuff in with fetchers
Keep api `getters` focussed on formatting response data and caching results
